requirements.txt
click
paho-mqtt
wakeonlan
apscheduler